2_youtube_summary/requirements.txt
pandas
pydantic
youtube-comment-downloader
langchain
langchain-core
langchain-openai
PyYAML
langdetect
